+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.45.0 (next)
 
+* Add `Connection::into_prepared` to allow for prepared queries to take ownership of the connection.
 * `Prepared::describe_param` now requires an exclusive reference
 * Remove `Prebound`.